items can have a quantity in inventory items can have a max quantity the player can hold, including 0 multiple items can be picked up/dropped/destroyed at once hash for locations, inventory and quests fixed

+++ services/aris/module.php
@@ -84,27 +84,65 @@ abstract class Module
     /**
      * Adds the specified item to the specified player.
      */
-     protected function giveItemToPlayer($strGamePrefix, $intItemID, $intPlayerID) {
-		    	
-    	$query = "INSERT INTO {$strGamePrefix}_player_items 
-										  (player_id, item_id) VALUES ($intPlayerID, $intItemID)
-										  ON duplicate KEY UPDATE item_id = $intItemID";
-		NetDebug::trace($query);
-		@mysql_query($query);
-		
+     protected function giveItemToPlayer($strGamePrefix, $intItemID, $intPlayerID, $qtyToGive=1) {
+		Module::adjustQtyForPlayerItem($strGamePrefix, $intItemID, $intPlayerID, $qtyToGive);
     }
 	
 	
 	/**
      * Removes the specified item from the user.
      */ 
-    protected function takeItemFromPlayer($strGamePrefix, $intItemID, $intPlayerID) {
-
-    	$query = "DELETE FROM {$strGamePrefix}_player_items 
-					WHERE player_id = $intPlayerID AND item_id = $intItemID";
-    	NetDebug::trace($query);
-    	@mysql_query($query);    	
+    protected function takeItemFromPlayer($strGamePrefix, $intItemID, $intPlayerID, $qtyToTake=1) {
+		Module::adjustQtyForPlayerItem($strGamePrefix, $intItemID, $intPlayerID, -$qtyToTake);
     }
+    
+    /**
+    * Updates the qty a player has of an item
+    */ 
+    protected function adjustQtyForPlayerItem($strGamePrefix, $intItemID, $intPlayerID, $amountOfAdjustment) {
+		
+		//Get any existing record
+		$query = "SELECT * FROM {$strGamePrefix}_player_items 
+					WHERE player_id = $intPlayerID AND item_id = $intItemID LIMIT 1";
+    	$result = @mysql_query($query);
+    	NetDebug::trace($query . mysql_error());
+
+    	if ($existingPlayerItem = @mysql_fetch_object($result)) {
+    		NetDebug::trace("We have an existing record for that player and item");
+
+ 			//Check if this change will make the qty go to < 1, if so delete the record
+ 			$newQty = $existingPlayerItem->qty + $amountOfAdjustment;
+ 			if ($newQty < 1) {
+ 				NetDebug::trace("Adjustment would result in a qty of $newQty so delete the record");
+ 				$query = "DELETE FROM {$strGamePrefix}_player_items 
+					WHERE player_id = $intPlayerID AND item_id = $intItemID";
+    			NetDebug::trace($query);
+    			@mysql_query($query);
+    		}
+    		else {
+ 				//Update the qty
+ 				NetDebug::trace("Updating Qty to $newQty");
+ 				$query = "UPDATE {$strGamePrefix}_player_items 
+ 							SET qty = $newQty
+							WHERE player_id = $intPlayerID AND item_id = $intItemID";
+    			NetDebug::trace($query);
+    			@mysql_query($query);
+ 			}
+    	}
+    	else if ($amountOfAdjustment > 0) {
+    		//Create a record
+    		NetDebug::trace("Creating a new player_item record");
+
+    		$query = "INSERT INTO {$strGamePrefix}_player_items 
+										  (player_id, item_id, qty) VALUES ($intPlayerID, $intItemID, $amountOfAdjustment)
+										  ON duplicate KEY UPDATE item_id = $intItemID";
+			NetDebug::trace($query);
+			@mysql_query($query);
+    	}
+    	else NetDebug::trace("Decrementing the qty of an item the player does not have. Ignored.");
+    	
+    }
+    
 	
 	/**
      * Decrement the item_qty at the specified location by the specified amount, default of 1
@@ -123,15 +161,49 @@ abstract class Module
      * Adds an item to Locations at the specified latitude, longitude
      */ 
     protected function giveItemToWorld($strGamePrefix, $intItemID, $floatLat, $floatLong, $intQty = 1) {
-		$itemName = $this->getItemName($strGamePrefix, $intItemID);
-		$error = 100; //Use 100 meters
-		$icon_media_id = $this->getItemIconMediaId($strGamePrefix, $intItemID); //Set the map icon = the item's icon
+		//Find any items on the map nearby
+		$clumpingRangeInMeters = 10;
 		
-		$query = "INSERT INTO {$strGamePrefix}_locations (name, type, type_id, icon_media_id, latitude, longitude, error, item_qty)
-										  VALUES ('{$itemName}','Item','{$intItemID}', '{$icon_media_id}', '{$floatLat}','{$floatLong}', '{$error}','{$intQty}')";
- 		NetDebug::trace($query);   	
-    	@mysql_query($query);    	
+		$query = "SELECT *,((ACOS(SIN($floatLat * PI() / 180) * SIN(latitude * PI() / 180) + 
+					COS($floatLat * PI() / 180) * COS(latitude * PI() / 180) * 
+					COS(($floatLong - longitude) * PI() / 180)) * 180 / PI()) * 60 * 1.1515) * 1609.344
+				AS `distance` FROM `{$strGamePrefix}_locations` HAVING `distance`<= {$clumpingRangeInMeters} ORDER BY `distance` ASC"; 	
+    	$result = @mysql_query($query);
+    	NetDebug::trace($query . mysql_error());  
+    	
+    	if ($closestLocationWithinClumpingRange = @mysql_fetch_object($result)) {
+    		//We have a match
+    		NetDebug::trace("An item exists nearby, adding to that location");   	
+
+    		$query = "UPDATE {$strGamePrefix}_locations
+    				SET item_qty = item_qty + {$intQty}";
+			NetDebug::trace($query);   	
+			@mysql_query($query);
+    	}
+		else {
+			NetDebug::trace("No item exists nearby, creating a new location");   	
+
+			$itemName = $this->getItemName($strGamePrefix, $intItemID);
+			$error = 100; //Use 100 meters
+			$icon_media_id = $this->getItemIconMediaId($strGamePrefix, $intItemID); //Set the map icon = the item's icon
+			
+			$query = "INSERT INTO {$strGamePrefix}_locations (name, type, type_id, icon_media_id, latitude, longitude, error, item_qty)
+											  VALUES ('{$itemName}','Item','{$intItemID}', '{$icon_media_id}', '{$floatLat}','{$floatLong}', '{$error}','{$intQty}')";
+			NetDebug::trace($query);   	
+			@mysql_query($query);
+    	}
     }
+	
+	protected function metersBetweenLatLngs($lat1, $lon1, $lat2, $lon2) { 
+
+		$theta = $lon1 - $lon2; 
+		$dist = sin(deg2rad($lat1)) * sin(deg2rad($lat2)) +  cos(deg2rad($lat1)) * cos(deg2rad($lat2)) * cos(deg2rad($theta)); 
+		$dist = acos($dist); 
+		$dist = rad2deg($dist); 
+		$miles = $dist * 60 * 1.1515;
+	 	$unit = strtoupper($unit);
+		return ($miles * 1609.344); //convert to meters
+	}
 	
     
     /**
@@ -370,12 +442,12 @@ abstract class Module
 			switch ($stateChange['action']) {
 				case Module::kPSC_GIVE_ITEM:
 					//echo 'Running a GIVE_ITEM';
-					Module::giveItemToPlayer($strPrefix, $stateChange['action_detail'], $intPlayerID);
+					Module::giveItemToPlayer($strPrefix, $stateChange['action_detail'], $intPlayerID,$stateChange['action_amount']);
 					$changeMade = TRUE;
 					break;
 				case Module::kPSC_TAKE_ITEM:
 					//echo 'Running a TAKE_ITEM';
-					Module::takeItemFromPlayer($strPrefix, $stateChange['action_detail'], $intPlayerID);
+					Module::takeItemFromPlayer($strPrefix, $stateChange['action_detail'], $intPlayerID,$stateChange['action_amount']);
 					$changeMade = TRUE;
 					break;
 			}
